resolve circular dependency
